project code
This program can be used by three distinct types of users:
1.	administrator
2.	player
3.	visitor
 The program has several functionalities assigned to each type of user mentioned above.
1.	ADMINISTRATOR
When the administrator enters the program, he/she will be prompted to enter the password to proceed further. Once the right password has been entered. He will be presented with a menu of tasks he can perform. The administrator can enter/delete players and book/cancel tickets for visitors at desired locations.

2.	PLAYER
The player can enter his/her ID and obtain details of the event including the score and rank he/she has obtained in the event.
 
3.	VISITOR
The visitor can check the status of his booking by entering the id which was assigned to him while he was booking the ticket and can also view the tickets that are available.


Overall this program efficiently implements several data structures to achieve this well-functioning system. This program can be used by event organizers to store details of players, get details of winners, book tickets for the visitor and maintain records of everything with real time update made to the data as it is altered ensuring atomicity. Hence this system is extremely useful and can be implemented a real-life scenario.